# backend/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "医院在线预约挂号系统 API",
    "description": "根据当前后端控制器、请求 DTO、测试断言与《项目分工与接口映射.md》整理的可导入 Apifox 的 OpenAPI 3.0 JSON。统一响应格式为 { code, message, data }。\n\n已补充 x-apifox-folder 和 operationId，导入后更容易按模块分组并生成接口结构。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发"
    }
  ],
...
  "paths": {
...
    "/api/hospitals": {
      "get": {
        "tags": [
          "医院"
        ],
        "summary": "医院列表",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "province",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageHospitalResponse"
                },
                "examples": {
                  "hospitalList": {
                    "summary": "医院列表",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "records": [
                          {
                            "id": 1,
                            "name": "北京第一人民医院",
                            "level": "三甲",
                            "province": "北京",
                            "city": "北京"
                          },
                          {
                            "id": 2,
                            "name": "上海瑞康医院",
                            "level": "三甲",
                            "province": "上海",
                            "city": "上海"
                          }
                        ],
                        "total": 2,
                        "page": 1,
                        "pageSize": 10
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "医院",
        "operationId": "getApiHospitals",
        "x-apifox-summary": "医院列表"
      }
    },
    "/api/hospitals/{id}": {
      "get": {
        "tags": [
          "医院"
        ],
        "summary": "医院详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HospitalDetailResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "医院",
        "operationId": "getApiHospitalsId",
        "x-apifox-summary": "医院详情"
      }
    },
    "/api/hospitals/{id}/departments": {
      "get": {
        "tags": [
          "医院"
        ],
        "summary": "医院科室树",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentTreeResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "医院",
        "operationId": "getApiHospitalsIdDepartments",
        "x-apifox-summary": "医院科室树"
      }
    },
    "/api/hospitals/{id}/doctors": {
      "get": {
        "tags": [
          "医院"
        ],
        "summary": "医院下的医生列表",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDoctorResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "医院",
        "operationId": "getApiHospitalsIdDoctors",
        "x-apifox-summary": "医院下的医生列表"
      }
    },
...
    "/api/messages": {
      "get": {
        "tags": [
          "消息"
        ],
        "summary": "消息列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageMessageResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "消息",
        "operationId": "getApiMessages",
        "x-apifox-summary": "消息列表"
      }
    },
    "/api/messages/{id}/read": {
      "post": {
        "tags": [
          "消息"
        ],
        "summary": "标记消息已读",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessNull"
          }
        },
        "x-apifox-folder": "消息",
        "operationId": "postApiMessagesIdRead",
        "x-apifox-summary": "标记消息已读"
      }
    },
    "/api/messages/unread-count": {
      "get": {
        "tags": [
          "消息"
        ],
        "summary": "未读消息数",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "消息",
        "operationId": "getApiMessagesUnreadCount",
        "x-apifox-summary": "未读消息数"
      }
    },
...
    "/api/follow/my": {
      "get": {
        "tags": [
          "关注"
        ],
        "summary": "我的关注列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "enum": [
                1,
                2,
                3
              ]
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageFollowResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "关注",
        "operationId": "getApiFollowMy",
        "x-apifox-summary": "我的关注列表"
      }
    },
    "/api/follow/{type}/{id}": {
      "post": {
        "tags": [
          "关注"
        ],
        "summary": "新增关注",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "enum": [
                1,
                2,
                3
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 4
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "createFollow": {
                    "summary": "关注成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "id": 4
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "关注",
        "operationId": "postApiFollowTypeId",
        "x-apifox-summary": "新增关注"
      },
      "delete": {
        "tags": [
          "关注"
        ],
        "summary": "取消关注",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "enum": [
                1,
                2,
                3
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessNull"
          }
        },
        "x-apifox-folder": "关注",
        "operationId": "deleteApiFollowTypeId",
        "x-apifox-summary": "取消关注"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        }
      }
    },
    "responses": {
      "SuccessNull": {
        "description": "成功",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "nullable": true,
                  "example": null
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "PageHospitalResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer",
                "example": 2
              },
              "page": {
                "type": "integer",
                "example": 1
              },
              "pageSize": {
                "type": "integer",
                "example": 10
              }
            }
          }
        }
      },
      "HospitalDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "北京第一人民医院"
              },
              "level": {
                "type": "string",
                "example": "三甲"
              },
              "address": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              },
              "intro": {
                "type": "string"
              },
              "image": {
                "type": "string"
              },
              "province": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "departmentCount": {
                "type": "integer"
              },
              "doctorCount": {
                "type": "integer"
              }
            }
          }
        }
      },
      "DepartmentTreeResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "children": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "PageDoctorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "PageMessageResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "PageFollowResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  }
}